refactor: complete MediaAssetStore resource seam
Close ADR-0006 by routing all media asset IO through MediaAssetStore, privatizing the filesystem implementation into internal adapters, and locking the boundary in pretest.
