+++ CONTEXT.md
@@ -64,6 +64,14 @@ _Avoid_：全局来源链接、全局站点评分
 影片刮削中的女优、男优、刮削标签和样张各自构成一个整体集合；更新时按字段整体写入或清空，不与旧刮削集合合并，其中手动标签不属于刮削标签集合。
 _Avoid_：增量合并、跨站点集合、手动标签覆盖
 
+**媒体资源**：
+位于媒体资源根目录内、由库拥有的本地图片文件（含明文与加密形态）；影片源视频文件不是媒体资源。
+_Avoid_：影片源文件、任意磁盘图片
+
+**媒体资源引用**：
+数据库中保存的、指向媒体资源的相对路径；引用本身不是文件操作能力。
+_Avoid_：绝对路径句柄、文件操作入口
+
 **影片图片字段写入**：
 影片封面和样张只有在所需图片全部成为可用本地资源后才整体写入；站点未返回图片属于空结果，返回图片但下载失败或文件不可读属于资源不可用，后者不得清空或部分替换已有资源。
 _Avoid_：下载失败清空、部分样张写入、失效图片路径
