+++ package.json
@@ -26,7 +26,8 @@
     "typecheck": "npm run typecheck:node && npm run typecheck:web",
     "check:encoding": "node scripts/check-encoding.mjs",
     "check:actress-boundaries": "node scripts/check-actress-boundaries.mjs",
-    "pretest": "npm run check:actress-boundaries",
+    "check:media-asset-store-boundaries": "node scripts/check-media-asset-store-boundaries.mjs",
+    "pretest": "npm run check:actress-boundaries && npm run check:media-asset-store-boundaries",
     "test": "npm run typecheck && node scripts/run-electron-tests.mjs",
     "mcp:plugin-dev": "tsx src/mcp/javdex-plugin-dev-server.ts",
     "rebuild": "electron-rebuild -f -w better-sqlite3",
