+++ src/main/services/actressGalleryService.ts
@@ -17,24 +17,21 @@ export async function importActressGalleryImage(
   if (input.source === 'file') {
     const sourcePath = input.sourcePath?.trim()
     if (!sourcePath) throw new Error('请选择本地图片文件')
-    const localPath = mediaAssetStore.importActressGallery(
-      actress.main_name,
-      sourcePath,
-      actressId
-    )
-    const dims =
-      mediaAssetStore.readStoredImageDimensions(localPath) ??
-      mediaAssetStore.readImageDimensionsAtPath(sourcePath)
-    try {
+    return mediaAssetStore.coordinateDatabaseChange(() => {
+      const localPath = mediaAssetStore.importActressGallery(
+        actress.main_name,
+        sourcePath,
+        actressId
+      )
+      const dims =
+        mediaAssetStore.readStoredImageDimensions(localPath) ??
+        mediaAssetStore.readImageDimensionsAtPath(sourcePath)
       return addActressGalleryAsset(actressId, {
         localPath,
         width: dims?.width ?? null,
         height: dims?.height ?? null
       })
-    } catch (error) {
-      mediaAssetStore.deleteBestEffort(localPath)
-      throw error
-    }
+    })
   }
 
   const rawUrl = input.remoteUrl?.trim()
@@ -50,28 +47,27 @@ export async function importActressGalleryImage(
   }
 
   const remoteUrl = parsed.toString()
-  const buf = await fetchRemoteImageBuffer(remoteUrl)
-  const downloaded = await mediaAssetStore.downloadActressGalleryImage(
-    actress.main_name,
-    remoteUrl,
-    async () => buf,
-    actressId
-  )
-  if (!downloaded) throw new Error('写真链接下载失败')
-  try {
+  return mediaAssetStore.coordinateDatabaseChangeAsync(async () => {
+    const buf = await fetchRemoteImageBuffer(remoteUrl)
+    const downloaded = await mediaAssetStore.downloadActressGalleryImage(
+      actress.main_name,
+      remoteUrl,
+      async () => buf,
+      actressId
+    )
+    if (!downloaded) throw new Error('写真链接下载失败')
     return addActressGalleryAsset(actressId, {
       remoteUrl,
       localPath: downloaded.localPath,
       width: downloaded.width,
       height: downloaded.height
     })
-  } catch (error) {
-    mediaAssetStore.deleteBestEffort(downloaded.localPath)
-    throw error
-  }
+  })
 }
 
 export function deleteActressGalleryImage(actressId: number, assetId: number): void {
-  const localPath = deleteActressGalleryAsset(actressId, assetId)
-  mediaAssetStore.deleteBestEffort(localPath)
+  mediaAssetStore.coordinateDatabaseChange(() => {
+    const localPath = deleteActressGalleryAsset(actressId, assetId)
+    mediaAssetStore.deleteBestEffort(localPath)
+  })
 }
